Handle complete set of data in table.
